Title: 바구니 안의 사탕, Time: 1071ms, Memory: 91MB, Status: Wrong Answer - CodeTree

+++ 240422/바구니 안의 사탕/candy-in-the-basket.java
@@ -1,7 +1,7 @@
 import java.util.*;
 
 public class Main {
-    public static long[] arr = new int[1000001];
+    public static long[] arr = new long[1000001];
     public static int maxLength;
     public static long maxValue = Long.MIN_VALUE;
